requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
numpy>=1.21.0
pandas>=1.3.0